web3_py_simple_storage/requirements.txt
web3
black
py-solc-x